+++ py/desiutil/svn/__init__.py
@@ -1,0 +1,14 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+# -*- coding: utf-8 -*-
+"""
+============
+desiutil.svn
+============
+
+This package contains code for interacting with DESI svn products.
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
+# The line above will help with 2to3 support.
+from .last_revision import last_revision
+from .last_tag import last_tag
+from .version import version