+++ bin/desiBootstrap.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# $Id$
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
 #
 function usage() {
     local execName=$(basename $0)